fix: report page XML links fixed to point to external page
Fixes issue https://github.com/IATI/validator-web/issues/862 where
links to XML files on the report page were built as an internal link
on the validator site.  Think this change in behaviour is due to
the removal of <StyledLink> which checked for external URLs first
and used :href instead of :to in those cases.  Changed this
<RouterLink> to set the link using :href as we know it's an external
link.